Address the comments of the first review round

+++ std/io.zig
@@ -232,9 +232,9 @@ pub fn OutStream(comptime WriteError: type) type {
             return self.writeFn(self, slice);
         }
 
-        pub fn writeByteNTimes(self: *Self, byte: u8, n: u64) Error!void {
+        pub fn writeByteNTimes(self: *Self, byte: u8, n: usize) Error!void {
             const slice = (*const [1]u8)(&byte)[0..];
-            var i: u64 = 0;
+            var i: usize = 0;
             while (i < n) : (i += 1) {
                 try self.writeFn(self, slice);
             }
